+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Scrap duplication.
 * Fixed bugs:
   * Duplicate scrap icon looking ugly.
+  * Failure decoding "therion.log". [reported by CaverBruce  (issue [#17](https://github.com/rsevero/mapiah/issues/17))]
 * Infrastructure maintenance:
   * Updated packaging/README.md to mention release constant.
   * Removed Flathub files from scripts/update_flutter_and_mapiah_version.dart.